refactor(attendances): move actions to controller
